Move configuration block into all_there.rb to avoid name clashes with requiring two different files called
